% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_2hin.R
\name{theme_2hin}
\alias{theme_2hin}
\title{A Custom ggplot2 Theme}
\usage{
theme_2hin(base_size = 12, base_family = "Times New Roman")
}
\arguments{
\item{base_size}{The base font size in points.}

\item{base_family}{The base font family. Defaults to "Times New Roman", which
is loaded with the package.}
}
\value{
A ggplot theme object (`theme`).
}
\description{
A complete ggplot2 theme using the Times New Roman font family, with a style
similar to `theme_grey`.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mpg, aes(displ, hwy, color = class)) +
  geom_point() +
  labs(title = "Fuel Economy by Engine Displacement") +
  theme_2hin()
}
}
